feat: roles, permissions, pws (#6)
* feat: roles, permissions, pws

* save work

* stuff

* fix tsconfig

* fix lint issues:

* move to individual pages

+++ components/templates/Users/UsersList/UsersList.tsx
@@ -1,20 +1,18 @@
 /* istanbul ignore file */
-import { Avatar, Badge, Table, Group, Text, ActionIcon, Anchor, ScrollArea, Select } from "@mantine/core";
+import Link from "next/link";
+import { Avatar, Badge, Table, Group, Text, ActionIcon, Anchor, ScrollArea } from "@mantine/core";
 import { IconPencil, IconTrash } from "@tabler/icons";
 
-import { ROLES } from "~constants";
 import { useFetchJson } from "~hooks";
 import { fetchJson } from "~utils";
-import type { User } from "~types";
 
 type UsersListProps = {
   fetchCount: number;
   setFetchCount: React.Dispatch<React.SetStateAction<number>>;
-  setUserToEdit: React.Dispatch<React.SetStateAction<User | null>>;
 };
 
 export default function UsersList(props: UsersListProps): JSX.Element {
-  const { fetchCount, setFetchCount, setUserToEdit } = props;
+  const { fetchCount, setFetchCount } = props;
 
   const res = useFetchJson("/api/users", {
     dependencies: [fetchCount],
@@ -22,6 +20,9 @@ export default function UsersList(props: UsersListProps): JSX.Element {
   const users = res.data;
 
   const handleDeleteUser = async (email: string): Promise<void> => {
+    if (!window.confirm(`Are you sure you want to delete the following user?\n\n${email}`)) {
+      return;
+    }
     await fetchJson("/api/users?type=delete", {
       method: "POST",
       body: JSON.stringify({ email }),
@@ -52,45 +53,47 @@ export default function UsersList(props: UsersListProps): JSX.Element {
         </thead>
         {users && Array.isArray(users) && users.length ? (
           <tbody>
-            {users.map((user) => (
-              <tr key={user.name}>
-                <td>
-                  <Group spacing="sm">
-                    <Avatar size={30} src={user.avatar} radius={30} />
-                    <Text size="sm" weight={500}>
-                      {user.name}
-                    </Text>
-                  </Group>
-                </td>
-                <td>
-                  <Anchor<"a"> size="sm" href="#" onClick={(event) => event.preventDefault()}>
-                    {user.email}
-                  </Anchor>
-                </td>
-                <td>
-                  <Select data={Object.values(ROLES)} defaultValue={user.role} variant="unstyled" />
-                </td>
-                <td>
-                  {Math.random() > 0.5 ? (
-                    <Badge fullWidth>Active</Badge>
-                  ) : (
-                    <Badge color="gray" fullWidth>
-                      Disabled
-                    </Badge>
-                  )}
-                </td>
-                <td>
-                  <Group spacing={0} position="right">
-                    <ActionIcon>
-                      <IconPencil size={18} stroke={1.5} onClick={() => setUserToEdit(user)} />
-                    </ActionIcon>
-                    <ActionIcon color="red" onClick={() => handleDeleteUser(user.email)}>
-                      <IconTrash size={18} stroke={1.5} />
-                    </ActionIcon>
-                  </Group>
-                </td>
-              </tr>
-            ))}
+            {users.map((user) => {
+              return (
+                <tr key={user.name}>
+                  <td>
+                    <Group spacing="sm">
+                      <Avatar size={30} src={user.avatar} radius={30} />
+                      <Text size="sm" weight={500}>
+                        {user.name}
+                      </Text>
+                    </Group>
+                  </td>
+                  <td>
+                    <Anchor<"a"> size="sm" href="#" onClick={(event) => event.preventDefault()}>
+                      {user.email}
+                    </Anchor>
+                  </td>
+                  <td>{user.role?.name}</td>
+                  <td>
+                    {user.disabled ? (
+                      <Badge color="gray" fullWidth>
+                        Disabled
+                      </Badge>
+                    ) : (
+                      <Badge fullWidth>Active</Badge>
+                    )}
+                  </td>
+                  <td>
+                    <Group spacing={0} position="right">
+                      <Link href={`/users/edit/${user.email}`}>
+                        <ActionIcon>
+                          <IconPencil size={18} stroke={1.5} />
+                        </ActionIcon>
+                      </Link>
+                      <ActionIcon color="red" onClick={() => handleDeleteUser(user.email)}>
+                        <IconTrash size={18} stroke={1.5} />
+                      </ActionIcon>
+                    </Group>
+                  </td>
+                </tr>
+              );
+            })}
           </tbody>
         ) : null}
       </Table>
